Fixed issue 858649 on Launchpad -- Distinguishes between tenant name and tenant id.
The recent changes to keystone switched from sending the tenant id to sending the tenant name as "tenant", thus breaking various tenant-related calls. This patch differentiates the two and names them as separate attributes on request.user. It also switches to using novaclient for the api.tenant_quota_get function.