More tenant novaclient conversions; Auth tenant list changes.
Converts most of the remaining tenant-related api functions to use novaclient instead of openstackx.

Changing tenant_list_for_token (which is used in the login process to choose a tenant) revealed a deeper bug wherein keystone might return a tenant that the user could administer but not log into as the first tenant, causing an Unauthorized message to be returned even though there were other tenants that could be logged into. This is now fixed.