Adds function to count tab characters in files to run_test.sh.
In case you're frustrated by mixed use of tab characters and
spaces for indentation, this function finds tab characters in
all the Python, CSS, Javascript and HTML files in the project,
and lists out any offending files that contain tabs.

Returns proper exit codes in case we someday want to gate on
this check.

Change-Id: I70195e9d00368a5d177cd3792921028e0bcd9b53