Unit tests shouldn't override template rendering.
Removes the fake_render_to_response function that was preventing
templates from being rendered. This required a number of adjustments
to the mocked obejcts in the view tests (which didn't need to be
mocked to begin with). A couple of legitimate problems were
exposed upon rendering the templates which are also now fixed.

Change-Id: I1647dcec01084318be1bb8f6d6ac1f4308992afa