Intelligent programmatic table generation!
Implements blueprint reusable-table-component.

This is a major step in making Horizon a componentized,
reusable suite of tools for building awesome dashboards. It
allows people to define and work with tables representing
live data in Python, constructing all the appropriate actions,
data transformations, etc. once, and then rendering everywhere
consistently and without all the copy-paste duplication.

Includes extensive tests and a full set of documentation for
all public parts of the API.

Change-Id: Ia4521eb1361a7fe5ebb839fde891db8fd5414c2e