Centralized error handling.
Implements blueprint exception-handling.

This takes care of a first-pass implementation at reusable,
consistent, centralized error handling.

It is known to be incomplete as far as the "recognized" exception
types. It needs to be expanded over time as it is further
put into real use and tested. This is only a starting point.

Change-Id: If19e7c1414456f1be69ad867995f46368749a9e9