Renamed all user-facing cases of "tenant" to "project".
This required a little more refactoring than expected because I
wanted to make the url slug also say "project", which required
changing the panel slug, which in turn required changing the name of
the panel directory, and then it seemed silly not to make the template
directory match.

Fixed bug 909495.

Change-Id: I9da3b544b46f507a223a44c256d0008c8674ab3b