Cleanup of quantum dependencies and imports.
Note: python-gflags is still a dependency for the quantum client,
but the client's setup.py properly lists it as a dependency so
we don't need it in our pip-requires.

Given that we now detect quantum fully from the keystone service catalog,
there are no settings/configuration for it in horizon anymore, and there
are no special dependencies for it beyond the client, I'm
saying that this fixes bug 893466.

Change-Id: I2f31215974082fd354ab7fd274a6643010640fa8