Removed "networks" panel from Horizon.
While this functionality will eventually be restored to Horizon,
when it does it will be in a significantly revamped and improved
form. For the Essex release, the amount of work required to get it
where it needs to be is too great and so it has regrettably been
cut from the current lineup of panels. :-(

Change-Id: I9c8b6e31303461a35270bea8d1ed30ad0ae6ed93