Instantiate novaclient managers for test data.
Some novaclient resources now have a "human_id" property
which requires a call to a bound method on the manager. Thereby
we need to instantiate the managers in the test data rather than
simply passing the class.

Fixes bug 958857.

There's also some cleanup on the test settings here, to remove
some outdated settings and to use an in-memory database instead
of a file on disk.

Change-Id: I1331e4ae5658b45d01b8247623554c5c43f67578