Paring down the syspanel instances table.
There was simply too much information in that table to be usable.
It probably still needs to lose more weight in favor of moving things
to a more admin-oriented detail page. However, we need to get real user
feedback on what data is critical for the overview vs. what can be in
a detail page.

Fixes bug 944506.

Change-Id: Iaad218bc6ed6647645c5943920467bec097a0a2d