Implements AJAX form posting.
This is somewhat of a hack for Essex, since the long-term solution
is a reworking of the way AJAX is handled. But it solves problems
in the interim and provides a significantly better experience.

Thanks to Andy Chong for pushing forward with the initial attempts
which lead to this patch.

Fixes bug 943518.

Change-Id: Ia65d926d3d406b07301e23b4c87de60c66ddec75