Add preference towards single id for action handling.
When the action handler received both a single id and multiple
object ids (such as when a user checks multiple boxes, then selects
a single row's action), it would previously favor the mutiple
ids, which goes against the intention of the user. This patch
causes it to favor the single id now.

Fixes bug 960866.

Change-Id: I53034589ada5792c22c2fc5144af2aee3b551eb0