volume.volumeId property used wrongly as .volume_id
Fixes bug 969162

Also, the created_at property was used as createdAt,
which lead to object reloading during template render
due to missing prop. See bug desc for more details.
Also, in the volume details template, the url link
for instance details was wrong - it was the link for
volume details instead.

Fixes Bug 971150

Change-Id: If5cfd515d0726a9ab85021b904a53e94b80205ea