Added a check for attachment['instance']
The problem, described in the ticket occurs
while you try to attach a volume to instance,
while at the same time terminating the instance(
I've succeeded to reproduce it that way).
As seen from the stack trace - the error comes
after an AJAX call for updating the table rows.
In normal situations with loading the page,
this error should not appear. Anyway, now there
is a check to prevent this error.

Fixes bug 999502

Change-Id: Ifccc2cf724fe14098c74b55e65f8fbdd8c5434dc