Make Horizon timezone-aware.
This systematically replaces anyplace that deals with dates or
times in Horizon with Django's timezone-aware machinery, and
enables timezone support in settings.

The assumption is that the server time should *always* be UTC.

TO DO: Add a setting for allowing the user to change their preferred
timezone display and add timezone indicators anywhere times are
displayed to the user.

Implements blueprint timezones. Also fixes bug 927974.

Change-Id: I5e462ba86e64b97b46873a017f87f328acee1b1d