Volume Progress Bar & Fixes For Quota
When a volume creation exceed the allocation quota
a vague error message was returned that offered the
user no guidance as to what went wrong. This has
been fixed.

Fixes Bug #1012883

This change also abstracts the Quota javascript to
allow it to be used anywhere on the site for any
progress bars that are to be shown in the future.

Implements blueprint progress-bar-javascript

On top of this I have renamed all "can_haz" filters
in the code, as they are really sort of embarassing.

Lastly, I have added the ability to append JS
events to the window load when a modal is loaded as
a static page, OR when it is loaded as an AJAX
modal.

Change-Id: I4b0cefa160cafbbd07d4b0981f62febaed051871