Fixed logic for disabling form button
This logic should be applied *only* for btn-primary
submit buttons, and not for regular buttons, which are
all of the action button in tables. At the moment this
logic breaks the table row actions, as it sets
button.disabled=true, and the 'action' value of the button
is not submitted, hence no action is performed.

Change-Id: I8e9827975656c53c7f8baff6554de1103bf4a358