Added action for creating a volume from snapshot
Added a new action to create new volume from
a volume snapshot. The snapshot_id is passed to
nova volume API. The 'Create Volume' form is
reused, with the 'size' value being checked against
the snapshot size, as it cannot be less than
the snapshot size.

Fixes bug 1023740

PATCH SET 4: Added dropdown to select snapshot source
in 'Create Volume' form, when loaded from /nova/volumes/
When a snapshot is selected from the dropdown, the name
and size are pre-filled in the form (I've removed
the description copy as it is probably specific
for the reason a snapshot is being created).

PATCH SET 6: rebased with master, fixed a bug with
get_context in nova/volume/views

Change-Id: I20dd8d698f5d8481938417557e09dcdc9b891e82