Update python-glanceclient to allow anything < v2
The python-glanceclient pythonic API shouldn't break until v2 is
released, so let's allow anything up to that known point.

Alters the way test exceptions are constructed so that we're
not dependent on the internals of how 3rd party libraries
handle their exceptions while still being able to use them
for real testing in Horizon.

Fixes bug 1035511

Change-Id: I9a4379a19d5d6dee9d2349a73d9321def45ad78d