Add nosehtmloutput as a test dependency.
Adding nosehtmloutput as a test dependency allows nose to output its
results to an html file. This will be used by Jenkins to save logs on
a different server.

Make slight modification to run_tests.sh in order to output the nose
test results to two different html files. One for Horizon and one for
the dashboard.

Change-Id: Ic1ed086b24705d0f2bca07f8e921646532e46c3b