User shouldn't remove their own roles on project
Added a check if the user is trying to delete their
own 'admin' role on the current project. This is added
to prevent the users from locking their own
access to Dashboard. Also, if the 'admin' role on the
current project is removed, Dashboard automatically
switches to another project for which the user has 'admin'
role, but the previous project still is displayed as current.

Fixes bug 1046538

Also improved the unit test scenarios a bit.

Change-Id: I1e60a11d628d6490ad24a8149b43ac307afb4780