Dynamically size the noVNC iframe, fix typos in copy
The iframe for the noVNC console is too small for some users who
like to use GUI on the console.  Leaving the sizes off completely
and letting the browser figure it out causes hilarity to ensue, so
explicitly set the height and width to quasi-sensible values.

Also, the "unavailable" error says "unavailabe"  Fixed.

Fixes Bug 1050316 ?

Addendum:

add an identifier to the iframe rather than using the auto-generated selector

Change-Id: I17cfa97cdc69f4b58e18d02451da2197c663e0d3