Allow permissions to be set on tables
Sometimes we want to hide certain UI elements depending on the
permissions granted to the current user. At the moment this is
possible for Actions and Panels. This commit extends support to
DataTable too.

We also set the permission on the volume snapshots table so that,
like the volumes panel, it is only displayed when the 'volume'
service is active.

Fixes bug #1087128.

Change-Id: Icc12b479c3eb888320af735b8b7810e58517eef0