Don't update count on tables without a summary row
The horizon.tabs javascript was updating the last row of any <table>
in the tab when the focus was changed to a different tab, stomping
over the last row of the table.

This commit limits updates to tables with the .datatable class,
and adds a class to the updated <span> in affected templates
so it can be selected a little more unambiguously.

Fixes bug #1080513.

Change-Id: I1b778bdc77887dcf540c96c2ab8be239bc13258c