Fixing sidebar layout in IE.
This problem is only reproducible when more than one panel group is
present in a dashboard (e.g., when Object Store is enabled in the
Project dashboard).  A crude way to reproduce the problem and verify
a fix is to comment out the "permissions" line in
openstack_dashboard/dashboards/project/containers/panel.py to enable
the second panel group.

The fix required was to remove an unnecessary float:left in
horizon.less that caused problems in IE's layout engine.  The
float:left had no effect on layout in Chrome, Firefox or
Safari(Windows version) in or out.

Fixes: bug #1078735
Change-Id: Idb0f8673d68ba686e1d7440fd2396e9a29b024ef