Replaces ugettext with ugettext_lazy in the dashboard-related files
While most files correctly import ugettext_lazy, some are importing ugettext instead.
This is causing some parts of the dashboard to not translate correctly when a user
switches languages. The core cause of this is because ugettext will translate the
string once when the django application starts. ugettext_lazy will translate when the
string is retrieved.

Change-Id: I44a394f4cda3b0f019e5a4601d4fa44fc110f939
Fixes: bug #1124276