Replaces ugettext with ugettext_lazy in the dashboard-related files
While most files correctly import ugettext_lazy, some are importing ugettext instead.
This is causing some parts of the dashboard to not translate correctly when a user
switches languages. The core cause of this is because ugettext will translate the
string once when the django application starts. ugettext_lazy will translate when the
string is retrieved.

Change-Id: Ibbdef60b3616d13aa7c2c5449ca592bf214a5f71
Fixes: bug #1124276