pin django to 1.4.x stream
django 1.5 just released in the last hour, and there is a
compatibility issue with it and horizon causing gate failures.

Pin this so that we can get back to business and let the horizon
team address django 1.5 when it makes sense to them.

Change-Id: I6ddf9623f8a4719b1cb82b118c1876373051fd28
(cherry picked from commit 0d53dc7063003fa7f02c030fd1631e0cd553d22a)