Don't change dashboard on admin tenant switch
Switching projects as an admin user currently dumps you back to
the Admin dashboard tab. This change fixes that by adding a 'next'
parameter to the switching URL that points to the default panel
of the current dashboard.

The redirect is to the default panel rather than the current one,
since the next tenant may not have access to the current panel
and we want to avoid kicking a user back to the login screen.

Fixes bug #1072848.

Change-Id: I25b8c0e38945d5db57f5bfe7bb174cdbd66156d5