Fixes error message formatting in Load Balancers panel
Strings were formatted to contain a "%s" to be updated when
the message was used.

The resulting string in the message box was:
    Error: Unable to add Pool "%s".

The update routine that already existed and was not being called.
Calls to the existing routine are now being made.  Other places
the strings were changed as the name would not be in context any
longer and the unused methods removed.

Fixes: bug #1152719

Change-Id: I157531a1b65dc73ce4ff1a57079fb4a2fcb9f62e