Fix adding security groups to an existing instance
Commit 4c34f5f1e8c66e4a865cf7996b283509feb91b71 introduced a
regression making it impossible to add a security group to
an existing instance.

In the javascript membership module, make sure new members are still
assigned the default role even if the role dropdown is not displayed
(as is the case for the instance security group add/remove view).

Fixes bug 1158037.

Change-Id: I986f7296e97a08d5e9141fe0739d4e028e066fc1