Remove inheritance of More action on first action
More indicator in the action dropdown should not inherit the class
of the first action. Any custom behavior attached to the first action
will also be inherited by More.

Example: If the first action had btn-danger, the More button become
red too. This might give the user the impression that More is also
a dangerous operation.

Fixes bug 1158419

Change-Id: Id078b738ba635ec47d2f32cde1917295f3ca5a79