Increase socket timeout for selenium tests
I can reproduce bug 1100771 consistently in my development environment.
Increasing the socket timeout in the SeleniumTestCase class fixes
the timeout issue for me.

This uncovers another issue that causes an "Address already in use
error". This happens because the order of tear down is wrong - we
need to quit() the driver before continuing with tear down rather
than after.

Fixes bug 1100771.

Change-Id: I95fbea05c3fc2516ce4e736eea9b0823bb5bbf79