Fix floating IP management from instance panel.
Quantum Floating IP API works with ports ids not instances ones
(opposite to what Nova Network does). So if we want to associate
(disassociate) an allocated floating IP to some instance we must
use the "target_id" (which is a port_id in case of Quantum and an
instance_id in case of Nova network).

Fixes bug 1158596.

Change-Id: Iceba4d8ee74bd58dc1cdbda7668841b0a7eb86f1