Fixes sorting on IP Address columns in tables.
The jQuery tablesorter plugin was choosing the digit parser
instead of the ipAddress parser for the IP Address columns.
This caused those columns to not sort correctly since only the
first 2 octects were looked at for sorting.

This patch adds an 'ip' data attribute to these columns and
updates the set_table_sorting function to set the ipAddress
sorter when this attribute is present.

Fixes bug 1108085

Change-Id: I3ca05048ce43b5bda14a4c586146c7ec9e479a79