Fix regression on running manage command.
Options on the manage command (./run_tests.sh) is getting shuffled,
this is due to the logic how testopts and testargs are extracted.

For example, the command "./run_tests.sh -m startdashboard test
--target test" is being translated into "startdashboard --target
test test".

The options after the "-m" should be passed as is as argument to the
manage command.

Fixes bug 1164505.

Change-Id: Id1a0699050390014487459757b3bda9c7fcca405