Don't duplicate networks list on keypair import
In the Launch Instance modal, if the user clicks on the button to import
a keypair, the networks listed in the Networking tab get duplicated.

The import keypair button calls `workflow_init` which in turn calls
`generate_networklist_html` which appended the networks to the list, not
clearing it beforehand.

This makes sure both Selected Networks and Available Networks lists are
emptied before the networks are added.

Fixes bug 1170193

Change-Id: Icb750caa08673068c509af6c8df1ad57a2a30d93
Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>
(cherry picked from commit 0512a12d07900430e8b399dcc27cade77d390491)