Group CRUD and Management in Admin Dashboard.
Add support for CRUD on Group for admin users. It also includes
the user management capability for the group.

Keystone revokes the token when a user is added or removed from the
group. If the logon user is added/removed from the group, the user
will be redirected to the login page.

This feature is only exposed if the user explicitly set horizon
for keystone V3.

Implements blueprint admin-group-crud

Change-Id: I1b5456af80bcc35e9d16ac6ba6792e82704e76fd