Adds methods for [] & len into LazyURLPattern
The Django 404 page currently depends on being able to call len()
and index LazyURLPattern.  However, the special methods for doing
this (__len__() and __getitem__()) are not forwarded to the object
wrapped by SimpleLazyObject by default.  Thus, methods to foward
these special methods to the wrapped object have been implemented
in LazyURLPattern.

Fixes: bug 1183578
Change-Id: I133e97ed66180fe78cfe1b7fd9c86bac4171a994