Fix bug in port creation form
'network_name' is not a recognized Quantum attribute, so passing it
in the request to the Quantum Server causes an error. In the end this
is not required because the network is identified by its id. So
this patch removes it (as it's done in other cases as well).

Fixes bug #1187129

Change-Id: Ie18c7bd504f8c241d002d6a050b8bcc59ed9fc03