Make the multi-select button on the "Floating IPs" page work again
This was a result of an assumption made by table-sorter.js, which
was parroted by Horizon; that there will only be one row of table
headers in a <thead> section.  Horizon uses two rows: one for the
"table title" and one for the individual table headings.

With the default settings to tablesorter(), this yielded an off-by-one
error when applying custom sorters (or parsers, depending on which
nomenclature you like, table-sorter.js uses two) to the table.

it is only tickled when manually defining custom parsers for columns,
and has been known about for a while, as there was code to compensate
for the observed behaviour.

Altering the xpath to omit the extra headers (and telling tablesorter
to do the same) yields the expected results.  I would personally
prefer to chain the results through .not(), but tablesorter only
provides support for xpath

Change-Id: I3338dd0cafd8271de6bf5d980359ed62c752ed98
Fixes: Bug 1186914