Refactoring url_for to remove admin parameter
Removed unused admin parameter that was unused and removed tests
that were no longer necessary.

Added an option configuration setting SECONDARY_ENDPOINT_TYPE that
will be attempted if the OPENSTACK_ENDPOINT_TYPE does not exist
in the service catalog for the desired service.

The primary use case for this fix is in cloud configurations
where Keystone does not return all endpoint types for each
service, and only does so based on the user's privilege level.

Example use case would be set OPENSTACK_ENDPOINT_TYPE to 'adminURL'
and set SECONDARY_ENDPOINT_TYPE to 'publicURL'.  If adminURL is not
available to the user, then they get the publicURL back.

If SECONDARY_ENDPOINT_TYPE is not set in the settings, then the
current behavior is maintained.

Fixes: bug #1186379

Change-Id: Ieefb6ed5dd88e5c840ef6bad93ae87237a1b63f9