Refactoring of user assignment workflow.
The javascript code and template for project membership has been
renamed to horizon.membership.js. It has been modified to take
step_slug as an argument, this is used as the identifier for the
membership step variables - for supporting multiple membership step
in a single workflow.

Also, removed unused template, tables and views.

Change-Id: I8685a307006cabc7005def160875c844e24c60ef
Fixes bug 1208558