Sync translation message catalogs with the latest code
In Grizzly branch several strings have been cleaned up since
the release. This commit update message catalog based on the
current code. This allows translators to use the latest ones
and it is good to sync them with the code before the release.

Fixes bug #1207442

Note that this patch is directly posted since message strings
in master are different and will be updated at different timing.

Change-Id: I53cfb5be2f791f5e446fba5a838fd2512d994146