JSHint integration
JSHint helps enforcing coding standards for the javascript files similarly to
pep8. It is integrated with the existing qUnit tests with the qhint library.
This approach has the advantage of running every time the selenium test suite
runs and does not introduce dependency to Node.js.

Partially implements: blueprint js-coding-style-checks

Change-Id: I77e75e7cdd5fd3426b9b981c4e0df5d30ef7b196