Delete and launch devices on the topology view
This feature enables to delete and launch a instance or router on the network
topology view. So you can do basic actions on this view.
Also it enables to change view small or normal. You can see more networks and
devices in the small view.

implements bp editable-network-topology-view

fixes bug #1215683

Change-Id: Ie65d50d2a99f72696c8f10223f8430ad5f90b144