Always display the Domains panel for Keystone V3
The Domains panel is only displayed when the user explicitly set
the OPENSTACK_KEYSTONE_MULTIDOMAIN_SUPPORT to True and keystone
supports V3.

The change allows the Domain panel to be displayed even if the flag
OPENSTACK_KEYSTONE_MULTIDOMAIN_SUPPORT is False. In this case, the
CRUD actions will not be available in the panel.

If the local_settings.py.example is taken as is, Keystone V3 is
assumed to be available. The Domain panel should also be available.

Change-Id: Ifa81c6f106ba98a192a163aeb80f0ff92888d954
Fixes: bug 1211667