Fix the Domain context for Projects panel
Add display of Domain ID and Domain Name in the Create and Update
Projects page to give user a hint where the Project will be created.
For Keystone V2, the two fields will be hidden.

When the Domain Context is set, the Project will be created in that
Domain Context. Otherwise, the Project will be created in the Domain
of the logon user.

On project creation, the listing of available users and groups will
be based on the Domain where it will be created. While on project
update, the listing of available users and groups will be based on
the Domain of the Project.

Fixes bug 1215196

Change-Id: Ia2ab25956621d01c2173e08402241b7f60401e0f