Fix the Domain Context for Users panel
Add display of Domain ID and Domain Name in the Create and Update
Users page to give user a hint where the User will be created.
For Keystone V2, the two fields will be hidden.

When the Domain Context is set, the User will be created in that
Domain Context. Otherwise, the User will be created in the Domain
of the logon user.

On users creation, the listing of available projects will be based
on the Domain where it will be created.

Fixes bug 1215697

Change-Id: I780d11289ab73804d0d26ad71a07ece6b2da74ad