Don't fidget with the image/snapshot name/size if it's set
Principle of Least Surprise: don't "help" users by setting field
names if they've already set them to something.  So if they've set
the volume name already, leave it alone.  If they've set the size
to something at least as large as what the source image/snapshot
requires, also leave it alone.

This means that if users are flipping through the available
images/snapshots, the UI will auto-set the values at most once.
This seems to be a workable trade-off as far as usability goes.

Fixes Bug: 1215566

Change-Id: I02c5ec0df1895ecaa24843712b9346ac0e588f59