Update error messeages in Launch Instance
If no image is selected when launching an instance, we will get
an error message "There are no image sources available; you must
first create an image before attempting to launch an instance."
In most cases this message is displayed when a user forget to
select an image. This commit changes the error message more generic
to avoid unnecessary confusion.

Closes-Bug: #1225755

It also changes the menu to display "No images available" if
no images are available rather than "Selct image" with no choice.

Change-Id: Ie5dd0760349c8b030baa2b2465776d0fa6c1767f