Don't display non-tenant Floating IP in Floating IP project tab
Previously Floating IPs from all tenants are displayed when logged
in as a user with admin role and Neutron is enabled. Neutron API
returns resources from all tenants when the API is called with admin
role, so we need to call the API with additional filter parameter to
limit the results to a given tenant.

Conflicts:
	openstack_dashboard/api/neutron.py
(cherry picked from commit 3827a7e73ab1daf317e3fbf903a57c31f95bf293)

Closes-Bug: #1226224
Change-Id: I47ddf77cd3c5d4c6141a67cb32c22cc917b485e6