Fix excessive volume_get requests in snaps table
The volume snapshots table has a column that shows the associated volume
name and links to its volume detail page. This calls volume_get twice
for each snapshot, so loading the Images & Snapshots page takes forever
if you have a decent number of snapshots.

This commit instead gets the volume list at the same time as the
volume snapshot list, and stores the volume as an attribute on
the snapshot so it can be easily retrieved later.

Closes-bug: 1218259
Change-Id: I004e50aae51a6ed7549da41967bc42babfa6cb45