Fixing database instance launch workflow
Two fixes related to one broken workflow.

First, 1227918 is a lack of support for trove backends that don't
have volume support enabled. To fix this, the user can now enter
into 0 for the volume size to work around this.  This alters how
the value is passed troveclient. Working with trove team to get a
programmatic way to determine is volume support is enabled.

Second, 1228416 is broken javascript when trying to launch a
database instance due to a couple of factors. The template for the
help is mislinked to use the project/instance one. Two, the fields
in the flavor that trove returns are different than the fields that
nova returns for flavor. This causes an exception when trying to
generate the quota/usage graphs on the right hand side of the form.
Since the javascript is looking for hardcoded values, this is not a
trivial fix so the offensive code has been removed until it can be
revisited. Additionally, there is no guarantee that the nova backend
trove uses is the same as the one regular compute uses. Needs more
thought/work.

Combined bug fixes because they are closely related to the same
action.

Change-Id: Ia378896d1d52067ac136c55ece9b527f44dd646b
Closes-bug: #1227918
Closes-bug: #1228416