Don't insert "Any Availaibility Zone" on single-AZ configurations
In cases where there is one AZ configured, Horizon displays the
option "Any Availability Zone" as the default.  This is confusing,
since there is only one.

Since there is logic to handle the "zero AZs configured" case
already, this is in fact an off-by-one error.  Alter the logic to
also handle the "one AZ configured" case and leave the single valid
option selected.

Fixes bug: 1229935

Change-Id: I49319e9c309c1ed41fd55e9297b0b461d17c726d