Fix editing of nameless images
We can't assume image.name is a valid attribute as Glance
doesn't send it for images that don't have a name.

This is a targeted fix to handle two cases where we access the name.
Where the name is not available, this shows the ID instead.

In the future we may want to look at wrapping images in our own class
so we can clean up data as it is returned from the API.

Change-Id: Ida50ce74efafff287a0189dc90aa529e80bd5961
Closes-bug: 1223921