Handle case without session_persistence attr in LBaaS VIP
When no session persistence mechanism is used, session_persistence
attribute is not contained in a VIP dict returned from Neutron.
This commit add a logic to handle this case.

During the fix it turns out the current LBaaS Horizon support
does not consider a case of "No Session Persistence".
This commit fixes the following too:

* Display "None" in VIP detail (instead of "Type:" without a value)
* Add "No session persistence" choice to VIP creation workflow
  (Actually it replaces "Select session persistence" choice
   since "No session persistence" is the default value.)

This commit also fixes bugs related to LBaaS VIP:

* Add validation for Cookie Name for APP_COOKIE persistence
  in VIP update. (The same is done for VIP creation)
* Do not create a hyperlink if no VIP exist for a pool.

Change-Id: Icf2aa2454f5f6beb928368fb06823576bbe57f2f
Closes-Bug: #1231704