The use of the class variables
All horizon/tables/actions.py action classes were rewriten, so there
is standard inicialization of objects, through the parameters of
initializer not through the class variables.
The class variable 'name' stayed though, because it is used for
ordering of the actions in the table. This will be solved in different
bug.

The BaseActionMetaClass has been added, so actions can inherit from
each other and use the class variable DSL in the same time. It is
similar solution to DataTableMetaclass.

Also the bugs with unused or blocking class atributes and [] in
parameters of initializer were solved.

Change-Id: I1b38befe114c1b4781e661a5a127c3f87391e228
Fixes: bug 1182931, bug 1183698, bug 1182908