Better @memoized decorator
This version of the @memoized decorator only keeps weak references to
the function arguments in cache, which means that it doesn't keep
objects alive forever, and that it won't fill up all available memory
eventually, as the old version would -- it only keeps cache for the
objects that are actually in use.

Honestly, I'm not actually sure that the effect is worth it, as the
level of complexity of this code is quite high. I thought I would
submit it anyways, so that we can at least have a discussion.

Change-Id: I3a4fca3aba7c52102673f7fa5d127873c71516f8
Partial-Bug: #1098047