Add volume type encryption to horizon
This patch adds support to horizon for the volume type encryption
changes made in cinder (https://review.openstack.org/#/c/30974/) and
in python-cinderclient (https://review.openstack.org/#/c/39984/).

The modifications are made to the Volume Type table, and add the
Create Encryption action to the action column.  Also, an Encryption
column is added, which includes a link to a summary of the encryption
information for the volume type.

Change-Id: I7f8e6396f8a1819ec6cb07f47ef6ba65832130c8
Implements: blueprint integration-with-cinder-volume-encryption