Deleting statistics tables from resource usage page
The tables were interpreting the stats in a wrong way, for some stats
it's not even easily possible to get those stats. We need to go through
each statistic separately and document it in the Ceilometer docs first.

Then it's more likely that these table stats will be distributed in
various tables in dashboard in the form of sparklines so these pages
won't be needed.

Fixes bug 1249279

Change-Id: Ib8c110f5d66fae50df12f8f0beeb99c2b2d00bf6