Add member to LBaaS when no servers available always succeed
Adding a member to a LBaaS pool when there are no servers
available it shows a "Success: Added member(s)." message.
Also, when there are no servers available, the message says
"Click Add to cancel" which is wrong from the UX perspective.
This patch makes the members field required in order to avoid
succeeding when cero members are added and changes the message
to be more representative.

Change-Id: I5c68b2cc8c5e457aeadee11711d36c802410f08d
Closes-Bug: #1260359