Friendlier information for lbaas members creation
In order to improve the user experience when adding new members to a
lbass pool, the help information shown in the 'Add Member' dialog has
been completed. Now, it explains that all selected members get the same
weight and port, and that later update is possible to achieve the
expected configuration.
In addition, the 'Weight' column has been added to the Member summary
tab, in order to easily identify the results of the performed 'Add
Member' action.
The complete fix would implement a more efficient way (through custom
widgets, perhaps) to select weight and port per instance.

Change-Id: If9a3ea6d36741fb1ca1ae06fb2dd307a78f537de
Partial-Bug: #1234688