CSS to fix content appearing below side pane
Content normally shows up on the right side of the navigation pane.
However, a long dashboard name will cause the content to appear below
the navigation pane.

Currently, the content is set to have left padding of 250px. This is
not ideal as the side-navigation pane may take up more space. I have
included a css fix that will allow the content to float and remain next
to the navigation pane. This solution have been tested in all major
browsers including IE.

Fixed the expanding issue.

Change-Id: Ied5c6085c643eba72a54279ff2e57221ddac76df
Closes-Bug: #1267661