Hide column on multi_select tables if no actions available.
This patch will add a runtime check to hide/show checkbox columns
on multi_select tables.
Right know the behaviour is calculated once but does not take in account
the dynamic allowed call of each Table Action available.

The change will add 'hidden' class to the first column if table actions
are present but all of them returned False on allowed() call.

Change-Id: Ia4512abe3dbb5de4581673717eb518b2d3a50291
Closes-Bug: #1271194