Add ability to manage image custom properties
This adds a new action on the admin images table for viewing the
image custom properties. From this new panel an admin can create,
delete, or update the custom properties. The new setting
IMAGE_RESERVED_CUSTOM_PROPERTIES can be used to specify any custom
properties that should not show up on this panel.

Change-Id: I10a6049e8cc6a06f31497e17a66d7031cd049c68
Implements: blueprint manage-image-custom-properties