Fix missing option in the test settings.py file
The settings.py file in the test directory for
openstack_dashboard is missing the profile_support option in
the OPENSTACK_NEUTRON_NETWORK setting.

This needs to be updated so that the tests for the "router"
dashboard and cisco N1K will be run correctly.

This fix merely ensures the OPENSTACK_NEUTRON_NETWORK setting
is similar in both the settings files.

Closes bug: #1278547

Change-Id: I51190cbe8f40bc6bf16dd246627de909456bd3a2