Fixing issue with VXLAN network profile create
This encompases the fix for bug 1229808 which requires change in
arguments to the network_profile_create API to create overlay
profiles instead of vxlan profiles.
Additionally overlay profiles now have a new parameter called
profile sub_types which lets one specify which type of overlay
profile can be created. This adds the new parameter as well.

Closes-bug: #1229808

Change-Id: I5f5153533533b0d760c04da33abf831d95f06ea4