More clear message when detaching volume
When detaching a volume from instance, the warning message looks like:
You have selected "/dev/vdb on instance test". Please confirm your
selection. This action cannot be undone.
If we consider that not all user's will be familiar with linux they may
not understand that vda in this warning means boot disk.
So use "volume abc on instance test" like message instead of the origin
one.

Change-Id: If1f7a161dfbc71ed22d286c65520b1500fb3a729
Closes-bug: 1281679