Add sort indicator to table headers
Currently, it is not possible to know which column determines the sort
order for the table. This patch adds a directional indicator that let
users know which column(s) is being sorted and the direction of the
sort.

known issue: because sorting indicator is a background image, it will
appear below column name if the name is too long.

Change-Id: If7194728d4335a04e9e8a7becd0086e2d3dc4b67
Closes-Bug: #1279578