Fix unbound local error in instance DetailView
If an error is raised in the instance DetailView we assume that a
redirect will be raised. However, certain exception classes will not
result in a redirect when handled by the standard Horizon exception
handler. If the error handler doesn't re-raise some sort of error,
control will be returned to the statement after the except block and
some of the necessary local variables may be unbound.

This fix simply ensures that an errors result in a redirect so that
we never end-up in a situation where the instance variable is unbound.

Change-Id: Ica9a09c9236bb50206277425633d63a8e7387401
Closes-Bug: 1291021