Fixing filesizeformat test issue with Django 1.6
In Django 1.6, the filesizeformat filter now applies
django.utils.html.avoid_wrapping to the formatted result. The end
result is replacing " " chars with "\x0a" a non-breaking space.
The string comparision of the error message in the affected test
was failing.

Adding a version check to allow support for Django 1.4 and 1.5.

Partially implements: blueprint django-1point6
Change-Id: I50f4c56f969d93b436abee6258e2e0102fd95784