Allow redisplay of the workflow from its handle()
Let the handle() method of the workflow raise a ValidationError to
indicate, that the operation failed and the whole workflow has to be
redisplayed for the user to correct the data and try again. It is
assumed that the handle() method will inject the right error messages
in the right steps and fields of the workflow.

This patch also cleans up the logic of the code a little, to make it
easier to read, and replaces the "next" variable that conflicts with
a python build-in with a more descriptive name.

Change-Id: I2fddb5ae8abbff72b8bf8258001a6f61aef3d1ba
Implements: bp/redisplay-workflow-after-handle-error