Refactor code around Console support
Add a console module to handle the different kinds on remote consoles
supported (vnc, rdp, spice ) and refactor the code
to use console.get_console.

In the future, a matrix based on HyperVisor could be added
in case that only one type of console is supported.
Ex: HyperV <=> RDP

Change-Id: If448b7c9d953765b9b56ee14b39975d951ffd92c
Implements: blueprint refactor-console-support
Closes-Bug: #1287881