Fix unbound locals in IPAssociationWorkflow
If an error is raised in the AssociateIPAction we assume that a redirect
will be raised. However, certain exception classes will not result in a
redirect when handled by the standard Horizon exception handler. If the
error handler doesn't re-raise some sort of error, control will be
returned to the statement after the except block and some of the
necessary local variables may be unbound.

This fix simply ensures that all of the necessary locals are bound in
the event that an error does occur.

Change-Id: Ib8c92ab218457e51e521d36bb4778c1dc1025342
Closes-Bug: #1290583