Add missing trunk segment type option for profile create
The Cisco N1K plugin allows for 'trunk' type network profiles to
be created in addition to existing 'vlan' and 'overlay' types.
This fix is to add to the segment_type choices in forms.py to
include the missing 'trunk' choice. Additionally the description
message has been updated to make sure strings are uniform in all
locations. The help string for the multicast range has also been
updated to reflect a valid mcast IP and finally new unit tests
have been added.

Closes-bug: #1285298

Change-Id: Ifc6d2952f2e37d6757fe7a432cffdb8fe36ffc61