Update .tx/config with type of files
The Jenkins job for horizon-upstream-translation-update expects that
.tx/config is setup properly. .tx/config is the transifex configuration
file that is used for interaction with transifix.com for translation of
the strings.

If you run the commands from the slave-script, it will update .tx/config
with information about the type of the files (PO-files). This change is
not needed and will break the periodic import script for horizon (which
does not exist).

Note that the slave-script needs a fix to work as well, will fix it in
the config repository. Also, the import script needs to be done.

Reference for jenkins job:
https://jenkins.openstack.org/view/All/job/horizon-upstream-translation-update/lastBuild/

Reference to slave-script:
http://git.openstack.org/cgit/openstack-infra/config/tree/modules/jenkins/files/slave_scripts/upstream_translation_horizon.sh

Change-Id: I9fb6799c8555487941967c2102e1d89c2530b60b