Changes in assert message and select dropdown
Three small changes in pageobject.py:

1. Updating the assert message in the page verification method
(is_the_current_page) to go along with the page title partial matching
way we decided on (reminder: the _page_title field only stores the
significant part of the page's title, due to the fact that the site
branding is configurable).

2. select_dropdown and select_dropdown_by_value now receivs the
webelement itself as an attribute, as oppose to an locator, needed to
be located again (by find_element). These methods aren't yet in
use, but definitely will be ,in patches I'm currently working on.

3. fill_field_by_locator was erased since it is not needed.
fill_field_element method receives the web element itself instead
of passing a locator (which is a private field of it's relevant object)
to fill_field_by_locator, finding it and then calling fill_field_element.
This way, fill_field_element is now used directly.

https://wiki.openstack.org/wiki/Horizon/Testing/UI

Partially implements blueprint: selenium-integration-testing

Change-Id: I12ade631069244876f0b4357c21d312255629437