Add user and password parameters to _do_login
The user and password parameters are needed for more complex login
tests (and negative ones) I plan on writing, as well as to a password
change test that will be submitted soon.
The default values for both attributes is None, so the login test
already merged won't be affected.

I also modified the if statement - it now checks whether the user is
the admin, otherwise functions the same way for all users.

https://wiki.openstack.org/wiki/Horizon/Testing/UI

Partially implements blueprint: selenium-integration-testing

Change-Id: Id2987103ab0933c0d3b5eb75a25679a9a9bcca55