Add go_to_home_page method and two more properties
I added the brand property which is used by the go_to_home_page
method. This method will generally be used for tests.
In particular, it will be part of the TestCase class, which is planned
to be the base class for all tests (on top of helpers's BaseTestCase
class), except login tests.

In addition, I modified the css selector in the _user_indicator_locator
and made the user property from it.

https://wiki.openstack.org/wiki/Horizon/Testing/UI

Partially implements blueprint: selenium-integration-testing

Change-Id: Ic614c6a37ea04aeccf1a06312f4bcd20e5fe2c76