Flavor access list titles consistent capitalization.
On the flavor access tab of the edit flavor workflow, you
can select the projects that have access to the flavor.
Currently, the left list is "All Projects" and the right
list is "Selected projects". The "p" should be capitalized
in Selected projects.

This changes "Projects" in both list titles to be capitalized.

Change-Id: Ica2958b50bd4d5dad7c1237bb2450a1c6f3e6efe
Closes-Bug: #1311715