Adding Jobs and Job Executions panels for Sahara
The Jobs and Job Executions panels allow for a job
to be defined (as a set of executables and libraries)
and ran on a cluster with a given set of data sources
(input/output) and configuration.  These panels are
incorporated in the same change because the job panel
has functionality that expects the job executions
panel to be in place (after launching a job, the
success url is the job executions page).
This code was originally from:
https://github.com/openstack/sahara-dashboard

Change-Id: Idf5061c71995c2baf3b6fc01a65c84a1ca90af44
Partial-Implements: blueprint merge-sahara-dashboard
Co-Authored-By: Nikita Konovalov <nkonovalov@mirantis.com>
Co-Authored-By: Dmitry Mescheryakov <dmescheryakov@mirantis.com>